removed attribute listeners
Removed attribute listeners for now to focus on using pane.viewer's load property to initalize & refresh code.

(Will add them back in later, along with compatibility slotted text when a less redundant method is found & everything is working as intended so that updates don't need to be done twice until the base functionality is figured out).